refactor any2mochi erlang
