+++ tools/any2mochi/convert_erl_golden_test.go
@@ -5,9 +5,11 @@ package any2mochi
 import (
 	"path/filepath"
 	"testing"
+
+	erl "mochi/tools/any2mochi/x/erlang"
 )
 
 func TestConvertErl_Golden(t *testing.T) {
 	root := findRepoRoot(t)
-	runConvertGolden(t, filepath.Join(root, "tests/compiler/erl"), "*.erl.out", ConvertErlangFile, "erl", ".mochi", ".error")
+	runConvertGolden(t, filepath.Join(root, "tests/compiler/erl"), "*.erl.out", erl.ConvertFile, "erl", ".mochi", ".error")
 }
